+++ src/extension.ts
@@ -3,6 +3,10 @@ import * as fs from 'fs';
 import axios, { AxiosResponse } from 'axios';
 import os = require("os");
 import path = require('path');
+import { window, workspace } from 'vscode';
+import * as qs from 'querystring';
+import { error } from 'console';
+import { Headers } from 'node-fetch';
 
 let ACCESS = false;
 var SelectedProject = 356;
@@ -18,7 +22,7 @@ let ventanaNueva: vscode.WebviewView;
 let extensions = vscode.extensions.all;
 extensions = extensions.filter(extension => !extension.id.startsWith('vscode.'));
 extensions.forEach(ex =>{
-  if (ex.id == "101OBEX, CORP.101obex-api-extension") ACCESS = true;
+  if (ex.id == "101OBeXCorp.101obex-api-extension") ACCESS = true;
 })
 
 var API_FOLDER_ACTIVE = '';
@@ -66,6 +70,11 @@ var UnpublishedAPI_LIST: any[] = [];
 
 var API_NAMES_LIST: any[] = [];
 var API_ENDPOINT_LIST: any[] = [];
+
+
+function isNumeric(value: string) {
+    return /^-?\d+$/.test(value);
+}
 
 export function activate(context: vscode.ExtensionContext) {
 
@@ -193,6 +202,14 @@ export function activate(context: vscode.ExtensionContext) {
 
 	} catch {}
 
+
+
+	/////7
+
+
+
+	//////7
+
 	try{
 		context.subscriptions.push(
 			vscode.commands.registerCommand('101obex-api-extension-api-publisher.addAPI-api-publisher', async () =>
@@ -240,13 +257,14 @@ export function activate(context: vscode.ExtensionContext) {
 			context.subscriptions.push(
 				vscode.commands.registerCommand('101obex-api-extension-api-publisher.linkAPI-api-publisher', async () =>
 				{
-	
+					if (cloud!='http://45.77.2.228:3000' && cloud!='http://144.202.116.237:3000'){
 					// Nombre del API
 					let APIMethodS = '';
 					let APITypeS = '';
+					let apiProdSel = '';
 					let error = false;
 					let ApiName = await vscode.window.showInputBox({
-						placeHolder: "Name of the API Product",
+						placeHolder: "Name of the API Link",
 						validateInput: text => {
 						return text === text ? null : 'Not 123!';
 						
@@ -274,6 +292,18 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 					// Endpoint
+
+					var ListaPIProd:  any[]  = [];
+					APIProductsList.forEach((element: any) => {
+						ListaPIProd.push(element.name);
+					});
+
+					const apiProd = await vscode.window.showQuickPick(
+						ListaPIProd,
+						{ placeHolder: 'Select API Product to publish in it' }).then(async (apiProd)=>{
+							apiProdSel = apiProd;
+
+						});
 
 					let TargetEndpoint = await vscode.window.showInputBox({
 						placeHolder: "Target Endpoint",
@@ -309,7 +339,8 @@ export function activate(context: vscode.ExtensionContext) {
 							endpoint: PublishedEndpoint,
 							ambient: 'local',
 							avap_catalog_id: 0,
-							name: ApiName
+							name: ApiName,
+							api_product: apiProdSel
 						}
 						const reqConfig = {
 							headers: {
@@ -349,12 +380,61 @@ export function activate(context: vscode.ExtensionContext) {
 						'API Link aborted'
 					);
 				}
+
+			} else {
+
+				vscode.window.showErrorMessage(
+					'You cannot link external APIs in Collaborative Network'
+				);
+
+			}
+
 				}
 				)
 					);
 			} catch { }
 
+			context.subscriptions.push(
+				vscode.commands.registerCommand('101obex-api-extension-api-publisher.ConfigTests', async (e) => {
+				  // Create and show panel
+				  let apiSele = e.description
+				  let ambient = 'local';
 
+				  
+				   
+				  const panel = vscode.window.createWebviewPanel(
+					'101obex-api-extension-api-publisher',
+					'Exetucion unitary test configuration',
+					vscode.ViewColumn.One,
+					{ enableScripts: true }
+					
+				  );
+			
+				  const cloud_params = `${cloud}/publishing/get_api_paremeters?obex_project_id=${SelectedProject}&ambient=${ambient}&endpoint=${apiSele}`
+				axios.get( cloud_params)
+						
+				.then(async (response) => { 
+				  // And set its HTML content
+
+
+					axios.get(`${cloud}/unittest/config?endpoint=${apiSele}&obex_project_id=${SelectedProject}`).then(async (response2)=>{
+						console.log(response2.data.data.config);
+						let con = []
+						if (response2.data.data.config.length>0){
+							 con = response2.data.data.config[0].response_validation;
+						} else {con = [];}
+						let parameters=response.data.data.parameters;
+						let results=response.data.data.response;
+						
+						if (typeof con !== 'object'){
+							let cc = JSON.parse(con);
+							con = cc;
+						}
+						panel.webview.html = getWebviewContent(parameters,results, AccesToken, apiSele, SelectedProject, con, cloud);
+					})
+				});
+				})
+			);
 
 	ExtContext = context;
 
@@ -471,6 +551,20 @@ export function activate(context: vscode.ExtensionContext) {
 						ListaPIProd.push(element.name);
 					});
 
+
+					axios.get(`${cloud}/unittest/config?name=${apiSele}&obex_project_id=${SelectedProject}`).then(async (response2)=>{
+
+						console.log(response2);
+						let zx = response2.data.data.config;
+						if (zx.length === 0){
+							vscode.window.showErrorMessage("No unitary test configured for this API, please configure it.");
+							return;
+						}
+					});
+
+
+
+
 					if (UnpublishedAPI_LIST.includes(apiSele)){
 
 					const pageType = await vscode.window.showQuickPick(
@@ -481,7 +575,7 @@ export function activate(context: vscode.ExtensionContext) {
 							if (pageType!=undefined) {
 								//console.log(e);
 								vscode.window.showInformationMessage('Publishind API');
-								let commit_id_selected = e.label;
+								let commit_id_selected = e.tooltip.split('|')[2];
 								let APIClone = e.tooltip.split('|')[0];
 								let AmbientPublish = e.tooltip.split('|')[1];
 
@@ -492,7 +586,13 @@ export function activate(context: vscode.ExtensionContext) {
 								axios.post(/*url3*/`${cloud}/publishing/publish_virtual_api?developer_token=` + dataObj.id_token+`&obex_project_id=${SelectedProject}&api=${APIClone}&commit_id=${commit_id_selected}&ambient=${AmbientPublish}&product=${pageType}`, axiosConfig)
 								.then((response) => {
 									//console.log(response);
+
+									
 									vscode.window.showInformationMessage('API Published');
+									if (cloud=='http://45.77.2.228:3000' || cloud == 'http://144.202.116.237:3000'){
+										vscode.window.showInformationMessage('APIs Need at least 2 minutes to propagate over the Collaborative Network');
+									}
+
 									vscode.commands.executeCommand('101obex-api-extension-api-publisher.refreshEntry-api-publisher');
 									vscode.commands.executeCommand('101obex-api-extension-api-creation.refreshEntry-api-creator')
 								});
@@ -524,6 +624,110 @@ export function activate(context: vscode.ExtensionContext) {
 
 				});
 				
+
+
+				
+
+
+				vscode.commands.registerCommand(`101obex-api-extension-api-publisher.TestThisAPI`, async (e) => {
+
+					let apiSele = e.tooltip.toString().split('|')[0];
+					let ambient = e.label;
+					let ambient_port = ambient === 'Live' ? 80:(ambient === 'Test' ? 81 : 82);
+					 let url_test = `${cloud.split(':')[0]}:${cloud.split(':')[1]}:${ambient_port}${apiSele}`
+					 let header_test = {'headers':{'101obextoken':SelectedProjectToken}}
+					
+					
+		let cloud_loc = 'http://0.0.0.0:3000'
+
+		const cloud_params = `${cloud}/publishing/get_api_paremeters?obex_project_id=${SelectedProject}&ambient=${ambient}&endpoint=${apiSele}`
+		axios.get( cloud_params)
+				
+		.then(async (response) => { 
+		
+
+
+			let names=response.data.data.parameters;
+			//let names=['id_canal']
+			let nameValuePairs = [];
+
+				for (const name of names) {
+					const value = await vscode.window.showInputBox({
+						prompt: `Insert value for: ${name}`,
+						placeHolder: `Value for ${name}`
+					});
+
+					if (value !== undefined) {
+						nameValuePairs.push({ name, value });
+					}
+				}
+							
+				const queryParams = nameValuePairs.reduce((acc: any, pair) => {
+					acc[pair.name] = pair.value;
+					return acc;
+				}, {});
+
+				const queryString = qs.stringify(queryParams);
+
+		
+		axios.get( `${url_test}?${queryString}` , header_test)
+				
+					.then((response) => {
+
+						console.log(response);
+						const wok = workspace.rootPath;
+						const terminal = window.createTerminal({
+							cwd: wok,
+							hideFromUser: false,
+							name: 'response',
+						});
+						terminal.show();
+						setTimeout(() => {
+							terminal.sendText(`clear && echo '\n---------------\nURL: ${url_test}\n---------------\nAPI Project Dev Token: ${SelectedProjectToken}\n---------------\nExample: \n    curl \"${url_test}?${queryString}\" --header \"101obextoken: ${SelectedProjectToken}\" \n---------------\nRESULT: ${JSON.stringify(response.data)}\n---------------\n\n'`);
+							console.log(JSON.parse(response.data.result))
+							}, 1000);
+						
+
+					}).catch((error) =>{
+						console.log(error);
+					});
+
+
+
+				}).catch((error) =>{
+					axios.get( url_test , header_test)
+				
+					.then((response) => {
+
+						console.log(response);
+						const wok = workspace.rootPath;
+						const terminal = window.createTerminal({
+							cwd: wok,
+							hideFromUser: false,
+							name: 'response',
+						});
+						terminal.show();
+						setTimeout(() => {
+							terminal.sendText(`clear && echo '\n---------------\nURL: ${url_test}\n---------------\nAPI Project Dev Token: ${SelectedProjectToken}\n---------------\nExample: \n    curl \"${url_test}\" --header \"101obextoken: ${SelectedProjectToken}\" \n---------------\nRESULT: ${JSON.stringify(response.data)}\n---------------\n\n'`);
+							console.log(JSON.parse(response.data.result))
+							}, 1000);
+						
+
+					}).catch((error) =>{
+						console.log(error);
+					});
+				});
+				
+				;
+
+
+
+
+				});
+
+
+
+
 				
 				}
 			)
@@ -566,6 +770,7 @@ class TreeDataProviderAPICreator implements vscode.TreeDataProvider<TreeItem> {
 			var subresponses8: TreeItem[] = [];
 			element.ambients.forEach((subelement: any) => {
 				var tttt: TreeItem[] = [];
+				var ttttd: TreeItem[] = [];
 				
 
 				subelement.versions.forEach((subsubelement:any) =>{
@@ -574,19 +779,36 @@ class TreeDataProviderAPICreator implements vscode.TreeDataProvider<TreeItem> {
 					if (subsubelement.tests != undefined) {
 					subsubelement.tests.forEach((subsubsubelement:any)=>{
 					var tttfilesystem: TreeItem[] = [];
-				//	if (subsubsubelement.code!=undefined) {
-				//		let code_parsed = subsubsubelement.code.replace(/\\"/g,'\"');
-				//		
-				//		code_parsed = JSON.parse(code_parsed);
-				//		code_parsed.forEach((subfile:any)=>{
-						tttfilesystem.push(new TreeItem(`Description`,[new TreeItem(`Execution test`,undefined,`resultado`,'')],`description|${subsubsubelement.result}`,''));
-						tttfilesystem.push(new TreeItem(`Result`,[new TreeItem(`${subsubsubelement.result}`,undefined,`resultado`,'')],`result|${subsubsubelement.result}`,''));
-						if (subsubsubelement.result!='ok') resultTests = 'ko';
-			
-						//		})
-				//	}
+					
+					//	if (subsubsubelement.code!=undefined) {
+					//		let code_parsed = subsubsubelement.code.replace(/\\"/g,'\"');
+					//		
+					//		code_parsed = JSON.parse(code_parsed);
+					//		code_parsed.forEach((subfile:any)=>{
+						
+						let resultado_test = '';
+						if (typeof subsubsubelement.result === 'object'){
+							if (subsubsubelement.result.success == true) {resultado_test = '{success:true}';} else {resultTests = 'ko'}
 
-						ttt.push(new TreeItem(`${subsubsubelement.name}`, tttfilesystem,'commit',`${subsubsubelement.result}`));
+						} else {
+							
+						if (!subsubsubelement.result.toString().includes('{succes:true}') && !subsubsubelement.result.toString().includes('"success":true')) {
+							resultTests = 'ko';
+						} else {
+							resultado_test = '{success:true}';
+						}
+						}
+
+						//tttfilesystem.push(new TreeItem(`Description`,[new TreeItem(`Execution test`,undefined,`resultado`,'')],`description|${resultado_test}`,''));
+			//			tttfilesystem.push(
+			//				new TreeItem(/*`Result`,[new TreeItem(*/`${resultado_test}`/*,undefined,`resultado`,'')]*/)
+						
+			//			);
+						//		})
+					//	}
+
+						//ttt.push(new TreeItem(`${subsubsubelement.name}`, tttfilesystem,'commit',`${resultado_test}`));
+		//				ttt.push(new TreeItem(`${resultado_test}`))
 					}) }
 
 					if (subsubelement.version!= 'unpublished') {
@@ -598,9 +820,9 @@ class TreeDataProviderAPICreator implements vscode.TreeDataProvider<TreeItem> {
 						  version_modified = `v${vv}`;
 						  
 					  }
-					tttt.push(new TreeItem(`${version_modified}`, ttt,`version ${resultTests}`,`${subsubelement.v_commit}`));
+					ttttd.push(new TreeItem(`${version_modified}`, undefined,`version ${resultTests}`,`${subsubelement.v_commit}`));
 					} else {
-					tttt.push(new TreeItem(`${subsubelement.commit_id}`, undefined,`unpublished|${element.name}|${subelement.ambient}|${subsubelement.message}`,`${subsubelement.commit_id}`));
+					tttt.push(new TreeItem(`${subsubelement.commit_id}`, ttttd,`unpublished|${element.name}|${subelement.ambient}|${subsubelement.message}`,`${subsubelement.commit_id}`));
 					}
 				} )
 				let ambient_modified = subelement.ambient.charAt(0).toUpperCase() + subelement.ambient.slice(1);
@@ -622,7 +844,7 @@ class TreeDataProviderAPICreator implements vscode.TreeDataProvider<TreeItem> {
 					new TreeItem(
 						`${ambient_modified}`,
 						tttt,
-						`API ${ultimaversion.toString()!='unpublished' && ultimaversion.toString()!='v0' ? `Last\u2000published\u2000${ultimaversion.toString()}` :'Unpublished'}`,
+						`API ${ultimaversion.toString()!='unpublished' && ultimaversion.toString()!='v0' ? `Last\u2000published\u2000${ultimaversion.toString()}|${element.endpoint}` :'Unpublished'}`,
 						`API`)
 						);
 			});
@@ -654,7 +876,7 @@ class TreeDataProviderAPICreator implements vscode.TreeDataProvider<TreeItem> {
 			}
 			let name_efective = element.name.charAt(0).toUpperCase() + element.name.slice(1);
 			name_efective = name_efective.replace("Staging","Test")
-			let apiType = lastversion == 0 ? 'API_VIRTUAL_NOT_PUBLISHED' :'API_VIRTUAL'
+			let apiType = element.method!=undefined ? (`LINK|${element.api_product_id}`) :(lastversion == 0 ? 'API_VIRTUAL_NOT_PUBLISHED' :'API_VIRTUAL')
 			
 
 			API_ENDPOINT_LIST.push(element.endpoint);
@@ -672,7 +894,9 @@ class TreeDataProviderAPICreator implements vscode.TreeDataProvider<TreeItem> {
 			
 			if (apiProductSelection == 'uncatalogued' || apiProductSelection == 'undefined' ||apiProductSelection == '' ) {
 				apiCat.tooltip = apiCat.tooltip?.toString().split('|')[0]
-				uncategorized.push(apiCat)
+				if (apiCat.children?.length === 1 && apiCat.children[0].children?.length == 1
+					&& apiCat.children[0].children[0].label == 'Cant publish init commit'
+				) {} else  uncategorized.push(apiCat)
 				UnpublishedAPI_LIST.push(apiCat.label);
 			};
 		})
@@ -682,7 +906,7 @@ let proooo = [];
 if (response!=undefined) proooo = response.data.data.Products
 		apiProducts.push(new TreeItem(`Uncategorized`, uncategorized,``,``))
 
-		proooo.forEach((apiP: { name: any; }) =>{
+		proooo.forEach((apiP: { name: any; id: any }) =>{
 			var tttt1: TreeItem[] = [];
 			category.forEach(apiCat=>{
 				let apiProductSelection = apiCat.tooltip?.toString().split('|')[1];
@@ -690,6 +914,17 @@ if (response!=undefined) proooo = response.data.data.Products
 				if (apiProductSelection === apiP.name){
 					apiCat.tooltip = apiCat.tooltip?.toString().split('|')[0]
 					tttt1.push(apiCat);
+				}
+				if (isNumeric(apiProductSelection || '')){
+					if (apiP.id === parseInt(apiProductSelection || '0')){
+						apiCat.tooltip = apiCat.tooltip?.toString().split('|')[0]
+						tttt1.push(apiCat);
+					}
+					if (parseInt(apiProductSelection || '0') === 0) {
+						apiCat.tooltip = apiCat.tooltip?.toString().split('|')[0]
+						uncategorized.push(apiCat)
+						UnpublishedAPI_LIST.push(apiCat.label);
+					}
 				}
 
 			});
@@ -748,7 +983,7 @@ class TreeItem extends vscode.TreeItem {
 	  this.iconPath = this.description === 'DEVELOPERS' ? path.join(__filename, '..', '..', 'images', 'person.svg') :  this.iconPath
 	  this.description = this.description === 'DEVELOPERS' ? "" : this.description
 
-	  this.iconPath = this.tooltip !== undefined ? path.join(__filename, '..', '..', 'images', 'settings.svg') :  this.iconPath
+	  this.iconPath = this.tooltip !== undefined ? new vscode.ThemeIcon('extensions-view-icon')/*path.join(__filename, '..', '..', 'images', 'settings.svg')*/ :  this.iconPath
 	  
 
 	  
@@ -792,7 +1027,7 @@ class TreeItem extends vscode.TreeItem {
 		if (addsd == 'commit'){
 			this.contextValue = '-COMMITCONF'
 			this.description = '';
-			this.iconPath = this.tooltip == "ok" ? path.join(__filename, '..', '..', 'images', 'test_ok.svg') : path.join(__filename, '..', '..', 'images', 'test_ko.svg');
+			this.iconPath = this.tooltip?.toString().includes('success:true')  ? path.join(__filename, '..', '..', 'images', 'test_ok.svg') : path.join(__filename, '..', '..', 'images', 'test_ko.svg');
 		}
 
 		if (addsd1 == 'file'){
@@ -808,7 +1043,7 @@ class TreeItem extends vscode.TreeItem {
 		}
 
 		if (this.label === API_FOLDER_ACTIVE) {
-			this.iconPath = path.join(__filename, '..', '..', 'images', 'settings-green.svg')
+			this.iconPath = path.join(__filename, '..', '..', 'images', 'cubos.svg')
 		}
 
 		if (addsd1 == 'result'){
@@ -876,6 +1111,7 @@ class TreeItem extends vscode.TreeItem {
 		if (addsd == 'version'){
 			this.description = this.description?.split(' ')[1] ;
 			this.iconPath = this.description == 'ok' ? path.join(__filename, '..', '..', 'images', 'test_ok.svg') : path.join(__filename, '..', '..', 'images', 'test_ko.svg'); //path.join(__filename, '..', '..', 'images', 'version.svg');
+			if (this.description == 'ok' || this.description == 'ko') this.description = '';
 		}
 
 		if (addsd == 'API'){
@@ -911,10 +1147,22 @@ class TreeItem extends vscode.TreeItem {
 
 		if (this.tooltip === 'link' || addsd2 === 'link'){
 			this.iconPath = path.join(__filename, '..', '..', 'images', 'link-icon.svg')
+			if (addsd2 === 'link' ) this.description = addsd1
 		}
 
-		
+		if (addsd1?.includes('Last\u2000published')){
+			this.contextValue = 'API_Published'
+			this.tooltip = this.description?.toString().split("|")[1];
+			this.description = addsd1;
+		}
+	console.log("--");	
+	console.log(this.contextValue);
+	if (document?.toString().includes('/') && (this.contextValue === undefined || this.contextValue == 'API_NOT_PUBLISHED')){
+		this.contextValue = "API_CONF";
 	}
+	console.log(this.label);
+	}
+	
   }
 
   function Connectors(
@@ -1018,9 +1266,10 @@ class TreeItem extends vscode.TreeItem {
 							// vscode.commands.executeCommand('react-webview-creation.start-low_code');
 						} 
 					}
-					if (e.description?.toString() == 'test'){
+					if (e.description?.toString().includes('////')){
 						//thisProvider.
-						TestEvent(e.tooltip?.toString().split('|')[1].toString());
+						let gg = e.tooltip?.toString();
+						TestEvent(gg);
 					}
 				}
 				);
@@ -1150,6 +1399,7 @@ function getCurrentCloud(){
 
 	var rawdata = fs.readFileSync(os.homedir+'/.101obex/selectedcloud.json');
 	var objectdata = JSON.parse(rawdata.toString());
+	//objectdata.selected_cloud = 'http://0.0.0.0:3000';
 	return objectdata
 }
 
@@ -1190,9 +1440,461 @@ async function setTestData(url: string, params: Object, init = false){
 	}
   }
 
+
+
+  /////
+
+  function getWebviewContent(parameters: any, results: any, developer_token:any, endpoint: any, selected_project: any, configuration: any, cloud:any) {
+	let api_values_data: never[] = [];
+	let headers: never[] = [];
+	return `<!DOCTYPE html>
+		<style>
+	input:focus {
+		outline: none;
+		border: specify yours;
+		}
+	.loader {
+		margin-left: 10px;
+		border: 6px solid #f3f3f3; /* Light grey */
+		border-top: 6px solid #da9300; /* Blue */
+		border-radius: 50%;
+		width: 10px;
+		height: 10px;
+		animation: spin 2s linear infinite;
+	  }
+	  
+	  @keyframes spin {
+		0% { transform: rotate(0deg); }
+		100% { transform: rotate(360deg); }
+	  }
+
+
+	  .clasenueva {
+	  	padding-top: 3px; 
+	  	padding-bottom:3px; 
+	  	height:24px;
+	  	padding-left:5px; 
+	  	width:100%; 
+	  	color: lightgrey;
+	  	border-style: solid;  
+	  	background-color: transparent; 
+	  	border-width: 0.5px; 
+	  	border-color: grey;
+	  	border-collapse: separate
+
+	  }
+	  .clasenueva:focus{
+	  	border-color: #1373d5;
+	  	border-width: 1px;
+	  	border-collapse: separate;
+	  	outline-color: #1373d5;
+	  	  box-shadow: 0 0 0 2px #1373d5; /* Borde azul simulado */
+  			margin-left: -0px; 
+	  }
+
+	</style>
+    <html>
+	<body onload="createForm();" style="">
+	
+	<div style="display: flex;">
+	<div style="min-width: 20px;margin-right: 10px; margin-top: 34px;"><span style="font-size:31px; color: white;">Unit test for ${endpoint}</span></div> 
+	</div>
+
+
+	<div style="">
+	<div style="min-width: 20px;margin-right: 10px; margin-top: 20px; color:white">API Params </div> 
+	
+	<div style="min-width: 20px;margin-right: 10px; margin-top: 12px; color:grey; width: 54%;">API parameters are inputs (path, query, or body) sent with API requests to specify, filter, or modify the behavior and response of the API.</div> 
+	</div>
+
+
+
+	<div style="display: table-caption; width:95vw">
+	<form style="margin-top: 10px;" id="formed"></form>
+	</div>
+
+
+	<div style="">
+	<div style="min-width: 20px;margin-right: 10px; margin-top: 20px; color:white">API Response </div> 
+	
+	<div style="min-width: 20px;margin-right: 10px; margin-top: 12px; color:grey; width: 54%;">API responses are the data returned by the server after processing a request, typically containing status codes, headers, and the requested resource or error details.</div> 
+	</div>
+
+
+
+	<div style="display: table-caption; width:95vw">
+	<form style="margin-top: 10px;" id="formed2"></form>
+	</div>
+
+
+
+	<div style="display: flex; margin-top: 40px;"> 
+	<button style="width:100px; border-style: solid; background-color: #1373d5; border-color: #1373d5; color: white" onclick="test2();" id="Save">Save</button> <div id='loader_spinner' class="loader"></div> 
+	</div>
+
+
+	<div style="display: table-caption;">
+	<form style="margin-top: 10px;" id="response"></form>
+	</div>
+
+
+	<div style="display: table-caption; color: transparent;">
+	<form style="margin-top: 10px;" id="respuestas"></form>
+	</div>
+
+	<div style="display: table-caption; color: transparent;">
+	<form style="margin-top: 10px;" id="respuestas2"></form>
+	</div>
+
+
+		<div style="display: table-caption;">
+	<form style="margin-top: 10px;" id="alertas"></form>
+	</div>
+
+
+
+	</body>
+
+	</html>
+	<script>
+
+    async function test2 () {
+    	let response_label = document.getElementById("response");
+		//response_label.innerText = 'Performing Request...';
+		let llo = document.getElementById("loader_spinner");
+		llo.hidden = false;
+
+		llo.hidden = false;
+		
+		
+		let params = [];
+		let gg = '${parameters}';
+		let ggg = gg.split(',')
+
+		ggg.forEach(g=>{
+			let cc = {}
+			cc[g]='-'
+			params.push(cc)}
+			)
+
+
+
+		let results = [];
+		let vv = '${results}';
+		let vvv = vv.split(',')
+
+		vvv.forEach(v=>{
+			let dd = {}
+			dd[v]='-'
+			results.push(dd)}
+			)
+
+
+
+		var resultvalues = document.getElementById("respuestas");
+
+		var paramvalues = document.getElementById("respuestas2");
+	
+		console.log(paramvalues.innerText);
+		const settings = {
+        	method: 'POST',
+        	headers: {
+            	Accept: 'application/json',
+            	'Content-Type': 'application/json',
+  				},
+			body: JSON.stringify({
+                     developer_token: '${developer_token}',
+                     endpoint: '${endpoint}',
+					 obex_project_id: '${selected_project}',
+					 param_config: JSON.stringify({
+					 parameters: paramvalues.innerText,
+					 results: resultvalues.innerText
+					 })
+                    })
+        	}
+
+
+
+
+
+
+    	let response = await fetch("${cloud}/unittest/config/",settings);
+    	if (response.ok) { // si el HTTP-status es 200-299
+		  // obtener cuerpo de la respuesta (método debajo)
+		  let json = await response.json();
+		//  console.log(json)
+//		  response_label.innerText = JSON.stringify(json);
+		} else {
+		  alert("Error-HTTP: " + response.status);
+		} 
+		llo.hidden = true;
+    };
+
+	function createForm(){
+		let llo = document.getElementById("loader_spinner");
+		llo.hidden = true;
+		let datat = {}
+		let datad = {}
+		let arrr = '${parameters}'
+		let brrr = '${results}'
+		//arrr = 'uno,dos'
+		//arrr = 'id_canal,codigo_pais'
+		//brrr = 'message'
+		let yy = '${api_values_data}';
+
+
+		arrrr = arrr.split(',');
+
+		brrrr = brrr.split(',');
+
+		if (yy!=undefined) yrrrr = yy.split(',')
+		
+		
+
+
+		let conta = 1;
+		for (i in arrrr) {
+
+			var entrada = document.createElement('div');
+				entrada.style.cssText = 'display: flex';
+
+
+		  form = document.getElementById("formed");
+	  
+		  var x = arrrr[i]
+		  
+		  
+		  const inputHandler = function(e) {
+
+
+
+			  
+		  datad[e.target.id] = e.target.value;
+
+		  let gfd = document.getElementById("respuestas2");
+		  gfd.innerText = JSON.stringify(datad);
+		  
+		  //let paramss= '?'
+		  //for (v in arrrr){
+			//  paramss+=arrrr[v]+'='+datat[arrrr[v]]+'&';
+		  //}
+		  //let req_urls = document.getElementById('req_url');
+		  //req_urls.value = '${url}'+paramss.substring(0,paramss.length-1);
+		  
+		}
+
+
+		
+		var input = document.createElement('input');  
+		input.setAttribute('id', x);
+
+		try{
+			let jjjjj1 = '${configuration.parameters}'
+			let jjjjjj1 = JSON.parse(jjjjj1)
+			if (jjjjjj1[x]!=undefined) input.value = jjjjjj1[x];
+  		} catch{}
+
+		 // if (yrrrr[i] != undefined && yrrrr!=undefined) input.value = yrrrr[i];
+		  //input.style.cssText = 'padding-top:3px;padding-bottom:3px;height:24px;padding-left:5px; border-style: solid; width: 32.87pc; background-color: transparent; border-width: 0.5px; border-color: grey; color: white;';
+		 input.classList.add("clasenueva") 
+		  input.addEventListener('input', inputHandler);
+	  
+	  	
+
+	  	if (conta==1){
+
+
+	  	var entrada0 = document.createElement('div');
+				entrada0.style.cssText = "display: flex; width:100%";
+
+		  var cabe_para = document.createElement('div') 
+	  	  cabe_para.style.cssText = 'height:25px;padding-left:5px; width:2em; color: lightgrey; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const cab_newContent = document.createTextNode("Id");
+	  	  cabe_para.appendChild(cab_newContent);
+	  	  entrada0.appendChild(cabe_para)
+	  	 
+
+
+		  var cabe_identi = document.createElement('div')
+	  	  cabe_identi.style.cssText = 'height:25px;padding-left:5px; width:20em; color: lightgrey; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const cabe_newContent2 = document.createTextNode("Param");
+	  	  cabe_identi.appendChild(cabe_newContent2);
+	  	  entrada0.appendChild(cabe_identi)
+
+
+	  	  var cabe_identi2 = document.createElement('div')
+	  	  cabe_identi2.style.cssText = 'height:25px;padding-left:5px; width:100%; color: lightgrey; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const cabe_newContent22 = document.createTextNode("Value");
+	  	  cabe_identi2.appendChild(cabe_newContent22);
+	  	  entrada0.appendChild(cabe_identi2)
+
+
+
+	  	  form.appendChild(entrada0);
+
+}
+
+
+
+
+	  	  var para = document.createElement('div')
+	  	  para.style.cssText = 'height:25px;padding-left:5px; width:20em; color: #878787; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const newContent = document.createTextNode(x);
+	  	  para.appendChild(newContent);
+
+
+		  var identi = document.createElement('div')
+	  	  identi.style.cssText = 'height:25px;padding-left:5px; width:2em; color: #878787; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const newContent2 = document.createTextNode(conta);
+	  	  identi.appendChild(newContent2);
+
+		conta++;
+
+		  if (x=!null && x!='') {
+		  entrada.appendChild(identi);
+		  	entrada.appendChild(para);
+		  	entrada.appendChild(input);
+
+		  	form.appendChild(entrada)
+		  }
+		}
+
+
+
+
+
+
+
+		conta = 1;
+
+		for (i in brrrr) {
+
+			var entrada = document.createElement('div');
+				entrada.style.cssText = "display: flex; width:100%";
+
+
+		  form = document.getElementById("formed2");
+
+
+
+
+	  	if (conta==1){
+
+
+	  	var entrada0 = document.createElement('div');
+				entrada0.style.cssText = "display: flex; width:100%";
+
+		  var cabe_para = document.createElement('div') 
+	  	  cabe_para.style.cssText = 'height:25px;padding-left:5px; width:2em; color: lightgrey; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const cab_newContent = document.createTextNode("Id");
+	  	  cabe_para.appendChild(cab_newContent);
+	  	  entrada0.appendChild(cabe_para)
+	  	 
+
+
+		  var cabe_identi = document.createElement('div')
+	  	  cabe_identi.style.cssText = 'height:25px;padding-left:5px; width:20em; color: lightgrey; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const cabe_newContent2 = document.createTextNode("Param");
+	  	  cabe_identi.appendChild(cabe_newContent2);
+	  	  entrada0.appendChild(cabe_identi)
+
+
+	  	  var cabe_identi2 = document.createElement('div')
+	  	  cabe_identi2.style.cssText = 'height:25px;padding-left:5px; width:100%; color: lightgrey; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const cabe_newContent22 = document.createTextNode("Regex Validation");
+	  	  cabe_identi2.appendChild(cabe_newContent22);
+	  	  entrada0.appendChild(cabe_identi2)
+
+
+
+	  	  form.appendChild(entrada0);
+
+}
+
+
+	  
+		  var x = brrrr[i]
+		  
+		  
+		  const inputHandler = function(e) {
+			  
+		  datat[e.target.id] = e.target.value;
+		  
+		  let gfd = document.getElementById("respuestas");
+		  gfd.innerText = JSON.stringify(datat);
+
+		  //let paramss= '?'
+		  //for (v in arrrr){
+			//  paramss+=arrrr[v]+'='+datat[arrrr[v]]+'&';
+		  //}
+		  //let req_urls = document.getElementById('req_url');
+		  //req_urls.value = '${url}'+paramss.substring(0,paramss.length-1);
+		  
+		}
+
+		var input = document.createElement('input');
+		input.setAttribute('id', x);
+
+		try{
+			let jjjjj = '${configuration.results}'
+			let jjjjjj = JSON.parse(jjjjj)
+			if (jjjjjj[x]!=undefined) input.value = jjjjjj[x];
+		} catch {}
+
+		 // if (yrrrr[i] != undefined && yrrrr!=undefined) input.value = yrrrr[i];
+		  //input.style.cssText = 'padding-top: 3px; padding-bottom:3px; height:24px;padding-left:5px; width:32.87pc; color: lightgrey;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+		 input.classList.add("clasenueva") 
+		  input.addEventListener('input', inputHandler);
+	  
+	  	
+	  	  var para = document.createElement('div')
+	  	  para.style.cssText = 'height:25px;padding-left:5px; width:20em; color: #878787; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const newContent = document.createTextNode(x);
+	  	  para.appendChild(newContent);
+
+
+		  var identid = document.createElement('div')
+	  	  identid.style.cssText = 'height:25px;padding-left:5px; width:2em; color: #878787; padding-top: 5px;border-style: solid;  background-color: transparent; border-width: 0.5px; border-color: grey;';
+	  	  const newContent2d = document.createTextNode(conta);
+	  	  identid.appendChild(newContent2d);
+
+		conta++;
+
+
+
+		  if (x=!null && x!='') {
+		  	entrada.appendChild(identid)
+		  	entrada.appendChild(para);
+		  	entrada.appendChild(input);
+
+		  	form.appendChild(entrada)
+		  }
+		}
+
+
+
+
+
+
+/*
+		if (arrrr.length == 0 ) {
+			form = document.getElementById("formed");
+			form.visivility = 'hidden';
+		}
+		*/
+	  }
+	</script>`;
+  }
+
+
+
+
+  /////
+
 function TestEvent(url: any, init = false) {  
 	let url_config = `${cloud}/info_api_parameters?developer_token=`
 	let pamameters_config = `&id_service=${url}&obex_project_id=${SelectedProject}`;
+	console.log("TESTEVENT");
 	if ((url!=null && url!='') || init){
 	if (!init) {
 		try{
@@ -1212,4 +1914,5 @@ function TestEvent(url: any, init = false) {
 	setTestData("", [], init);
 }
 }
+
 }